+++ config/routes/collections.php
@@ -46,10 +46,10 @@ return function (App $app) {
 		$group->patch('/{collection}/{id}/{property}/{name}', Action\Object\ObjectPatchPropertyMetaAction::class)->setName('property-meta-patch');
 
 		// Deck Items (for deck properties)
-		$group->post('/{collection}/{id}/{property}/deck', Action\Object\Deck\DeckItemCreateAction::class)->setName('deck-item-create');
-		$group->get('/{collection}/{id}/{property}/deck/{itemId}', Action\Object\Deck\DeckItemFetchAction::class)->setName('deck-item-fetch');
-		$group->put('/{collection}/{id}/{property}/deck/{itemId}', Action\Object\Deck\DeckItemUpdateAction::class)->setName('deck-item-update');
-		$group->delete('/{collection}/{id}/{property}/deck/{itemId}', Action\Object\Deck\DeckItemDeleteAction::class)->setName('deck-item-delete');
+		$group->post('/{collection}/{id}/{property}/deck', Property\Deck\DeckItemCreateAction::class)->setName('deck-item-create');
+		$group->get('/{collection}/{id}/{property}/deck/{itemId}', Property\Deck\DeckItemFetchAction::class)->setName('deck-item-fetch');
+		$group->put('/{collection}/{id}/{property}/deck/{itemId}', Property\Deck\DeckItemUpdateAction::class)->setName('deck-item-update');
+		$group->delete('/{collection}/{id}/{property}/deck/{itemId}', Property\Deck\DeckItemDeleteAction::class)->setName('deck-item-delete');
 
 		// Property File
 		$group->post('/{collection}/{id}/{property}', Property\File\FileSaveAction::class)->setName('property-file-save');